Remove auth directory and rename globus_auth to auth
